+++ .codecov.yml
@@ -1,0 +1,1 @@
+comment: false
