+++ cmd/userfeatureadd.go
@@ -1,8 +1,7 @@
 package cmd
 
 import (
-	"github.com/nupplaphil/kopano-ldap/lib/kopano"
-	"github.com/nupplaphil/kopano-ldap/lib/utils"
+	"github.com/nupplaphil/kopano-ldap/kopano"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"log"
@@ -46,10 +45,14 @@ func runUserFeatureAdd(flags *pflag.FlagSet) {
 		os.Exit(1)
 	}
 
-	host, port, fqdn, ldap_user, password := LdapFlags()
-	baseDn := utils.GetBaseDN(fqdn)
+	ldapHost, ldapPort, ldapDomain, ldapUser, ldapPW := LdapFlags()
+	baseDn := kopano.GetBaseDN(ldapDomain)
 
-	client := kopano.Connect(host, port, fqdn, ldap_user, password)
+	client, err := kopano.Connect(ldapHost, ldapPort, ldapDomain, ldapUser, ldapPW)
+	if err != nil {
+		log.Fatal(err)
+		os.Exit(1)
+	}
 
 	kopano.AddUserFeatures(client, baseDn, user, features)
 }
