+++ cmd/userfeaturerem.go
@@ -1,8 +1,7 @@
 package cmd
 
 import (
-	"github.com/nupplaphil/kopano-ldap/lib/kopano"
-	"github.com/nupplaphil/kopano-ldap/lib/utils"
+	"github.com/nupplaphil/kopano-ldap/kopano"
 	"github.com/spf13/pflag"
 	"log"
 	"os"
@@ -48,10 +47,14 @@ func runUserFeatureRemove(flags *pflag.FlagSet) {
 		os.Exit(1)
 	}
 
-	host, port, fqdn, ldap_user, password := LdapFlags()
-	baseDn := utils.GetBaseDN(fqdn)
+	ldapHost, ldapPort, ldapDomain, ldapUser, ldapPW := LdapFlags()
+	baseDn := kopano.GetBaseDN(ldapDomain)
 
-	client := kopano.Connect(host, port, fqdn, ldap_user, password)
+	client, err := kopano.Connect(ldapHost, ldapPort, ldapDomain, ldapUser, ldapPW)
+	if err != nil {
+		log.Fatal(err)
+		os.Exit(1)
+	}
 
 	kopano.RemoveUserFeatures(client, baseDn, user, features)
 }
